extracted windowSelectionHelpers
